don't use C array as method parameter or return value
There are several places in ITK where C arrays are used as method
parameter or return value instead of stronger array like types, like
itk::Index, itk::Size, itk::Offset, itk::FixedArray or itk::Array.
Using stronger types makes more efficient the error checking at compile
time and enhance the user integration by using the same types and
conventions than elsewhere.

More important, the C arrays are not wrappable without further
information that we don't have at wrap time. As a consequence, the
classes which are using them are at best partially usable in Python,
Java or Tcl. Using stronger types completely fix that problem.

No effort has been made up to now to preserve backward compatibility.
There is some room for enhancement on that point, but backward
compatibility shouldn't be 100% achievable unless the affected method
are also renamed -- this may be possible for a few of them.

+++ Testing/Code/BasicFilters/itkVectorExpandImageFilterTest.cxx
@@ -143,7 +143,7 @@ int itkVectorExpandImageFilterTest(int, char* [] )
 
   expander->SetExpandFactors( 5 );
 
-  unsigned int factors[ImageDimension] = {2,3};
+  float factors[ImageDimension] = {2,3};
   expander->SetExpandFactors( factors );
 
   typedef ImageType::PixelType PixelType;
